Add provider_peering_id to peering resource
This commit adds a new computed field to the peering resource which
exposes the resourse-provider-assigned peering identifier. This allows
other Terraform code to query for the peering link and accept it in a
reliable fashion.

As part of this, we refactor the waiter pattern in the client to return
the resource representation from the response which moves into the
target state to avoid an additional network request. This should
ultimately be propagated to the network and cluster waiters also.