Fix peering update, allow deletes of defunct resources
There was some logic that would check to see if a peering resource was
in a defunct state, and send an error if it was. This was being called
from creates and updates, but also from the normal read which meant
it was impossible to destroy a peering resources that was already in
a defunct state.

This also fixes a bug in peering updates which prevented it from
exiting.