Adds support back for server version 20.6
The Terraform provider was changed awhile back to not allow 20.6, as
it was out of support. This was a good idea, but it introduces a problem
that old terraform files which mention clusters with server version 20.6
now raise validation errors that prevent otherwise innocuous actions
such as running "terraform plan" from working.

20.6 no longer works with the cluster create call. I'm going to tweak
the error message shown for the create call to let users know 20.6 was
supported but isn't anymore, which is then what they'll see if they try
to create a cluster with 20.6.